Move external archival functions to separate file
