+++ ledger/ledger.go
@@ -17,18 +17,12 @@
 package ledger
 
 import (
-	"bytes"
 	"context"
 	"database/sql"
 	"fmt"
-	"io"
-	"net"
-	"net/http"
-	"os"
 	"path/filepath"
 	"time"
 
-	"github.com/algorand/go-codec/codec"
 	"github.com/algorand/go-deadlock"
 
 	"github.com/algorand/go-algorand/agreement"
@@ -391,7 +385,7 @@ func initBlocksDB(tx *sql.Tx, log logging.Logger, initBlocks []bookkeeping.Block
 	//
 	// When running in external archival mode, this check is not needed
 	// (since those blocks will be available in external storage).
-	_, externalArchivalEnabled := externalArchivalSettings()
+	_, externalArchivalEnabled := externalArchiveSettings()
 	if externalArchivalEnabled {
 		log.Info("skipping block init check in external archival mode")
 		return nil
@@ -771,7 +765,7 @@ func (l *Ledger) LatestCommitted() (basics.Round, basics.Round) {
 func (l *Ledger) Block(rnd basics.Round) (blk bookkeeping.Block, err error) {
 
 	// Try to obtain the block from local storage.
-	url, externalArchivalEnabled := externalArchivalSettings()
+	url, externalArchivalEnabled := externalArchiveSettings()
 	blk, err = l.blockQ.getBlock(rnd)
 	if err == nil || !externalArchivalEnabled {
 		return blk, err
@@ -780,7 +774,7 @@ func (l *Ledger) Block(rnd basics.Round) (blk bookkeeping.Block, err error) {
 	// Fall back to external storage.
 	switch err.(type) {
 	case ledgercore.ErrNoEntry:
-		tmp, err := downloadEncodedBlockCertFromArchive(rnd, url)
+		tmp, err := downloadBlockFromExternalArchive(rnd, url)
 		return tmp.Block, err
 	default:
 		return blk, err
@@ -806,7 +800,7 @@ func (l *Ledger) BlockHdr(rnd basics.Round) (hdr bookkeeping.BlockHeader, err er
 	}
 
 	// Try to obtain the block header from local storage.
-	url, externalArchivalEnabled := externalArchivalSettings()
+	url, externalArchivalEnabled := externalArchiveSettings()
 	hdr, err = l.blockQ.getBlockHdr(rnd)
 	if err == nil || !externalArchivalEnabled {
 		return hdr, err
@@ -815,7 +809,7 @@ func (l *Ledger) BlockHdr(rnd basics.Round) (hdr bookkeeping.BlockHeader, err er
 	// Fall back to external storage.
 	switch err.(type) {
 	case ledgercore.ErrNoEntry:
-		tmp, err := downloadEncodedBlockCertFromArchive(rnd, url)
+		tmp, err := downloadBlockFromExternalArchive(rnd, url)
 		return tmp.Block.BlockHeader, err
 	default:
 		return hdr, err
@@ -826,7 +820,7 @@ func (l *Ledger) BlockHdr(rnd basics.Round) (hdr bookkeeping.BlockHeader, err er
 func (l *Ledger) EncodedBlockCert(rnd basics.Round) (blk []byte, cert []byte, err error) {
 
 	// Try to obtain the block from local storage.
-	url, externalArchivalEnabled := externalArchivalSettings()
+	url, externalArchivalEnabled := externalArchiveSettings()
 	blk, cert, err = l.blockQ.getEncodedBlockCert(rnd)
 	if err == nil || !externalArchivalEnabled {
 		return blk, cert, err
@@ -835,7 +829,7 @@ func (l *Ledger) EncodedBlockCert(rnd basics.Round) (blk []byte, cert []byte, er
 	// Fall back to external storage.
 	switch err.(type) {
 	case ledgercore.ErrNoEntry:
-		return downloadEncodedBlockCertFromExternalArchive(rnd, url)
+		return downloadBlockBytesFromExternalArchive(rnd, url)
 	default:
 		return blk, cert, err
 	}
@@ -845,7 +839,7 @@ func (l *Ledger) EncodedBlockCert(rnd basics.Round) (blk []byte, cert []byte, er
 func (l *Ledger) BlockCert(rnd basics.Round) (blk bookkeeping.Block, cert agreement.Certificate, err error) {
 
 	// Try to obtain the block from local storage.
-	url, externalArchivalEnabled := externalArchivalSettings()
+	url, externalArchivalEnabled := externalArchiveSettings()
 	blk, cert, err = l.blockQ.getBlockCert(rnd)
 	if err == nil || !externalArchivalEnabled {
 		return blk, cert, err
@@ -854,7 +848,7 @@ func (l *Ledger) BlockCert(rnd basics.Round) (blk bookkeeping.Block, cert agreem
 	// Fall back to external storage.
 	switch err.(type) {
 	case ledgercore.ErrNoEntry:
-		tmp, err := downloadEncodedBlockCertFromArchive(rnd, url)
+		tmp, err := downloadBlockFromExternalArchive(rnd, url)
 		return tmp.Block, tmp.Certificate, err
 	default:
 		return blk, cert, err
@@ -1083,112 +1077,3 @@ var ledgerVerifygenhashCount = metrics.NewCounter("ledger_verifygenhash_count", 
 var ledgerVerifygenhashMicros = metrics.NewCounter("ledger_verifygenhash_micros", "µs spent")
 var ledgerTrackerMuLockCount = metrics.NewCounter("ledger_lock_trackermu_count", "calls")
 var ledgerTrackerMuLockMicros = metrics.NewCounter("ledger_lock_trackermu_micros", "µs spent")
-
-// externalArchivalSettings indicates whether archive mode is enabled.
-//
-// In case of being enabled, the URL of the external archive is returned.
-// When disabled, this URL is set to "".
-func externalArchivalSettings() (url string, enabled bool) {
-	url = os.Getenv("EXTERNAL_ARCHIVE_URL")
-	return url, url != ""
-}
-
-func downloadEncodedBlockCertFromExternalArchive(rnd basics.Round, baseUrl string) (blk []byte, cert []byte, err error) {
-
-	encodedBlockCert, err := downloadEncodedBlockCertFromArchive(rnd, baseUrl)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	// Encode block bytes to msgpack
-	{
-		codecHandle := makeCodecHandle()
-
-		var buf bytes.Buffer
-		enc := codec.NewEncoder(&buf, codecHandle)
-		err = enc.Encode(encodedBlockCert.Block)
-		if err != nil {
-			return nil, nil, fmt.Errorf("failed to encode EncodedBlockCert.Block to msgpack: %w", err)
-		}
-		blk = buf.Bytes()
-	}
-
-	// Encode certificate bytes to msgpack
-	{
-		codecHandle := makeCodecHandle()
-
-		var buf bytes.Buffer
-		enc := codec.NewEncoder(&buf, codecHandle)
-		err = enc.Encode(encodedBlockCert.Certificate)
-		if err != nil {
-			return nil, nil, fmt.Errorf("failed to encode EncodedBlockCert.Certificate to msgpack: %w", err)
-		}
-		cert = buf.Bytes()
-	}
-
-	return blk, cert, nil
-}
-
-// EncodedBlockCert defines how GetBlockBytes encodes a block and its certificate
-type EncodedBlockCert struct {
-	_struct struct{} `codec:""`
-
-	Block       bookkeeping.Block     `codec:"block"`
-	Certificate agreement.Certificate `codec:"cert"`
-}
-
-func downloadEncodedBlockCertFromArchive(rnd basics.Round, baseUrl string) (blk EncodedBlockCert, err error) {
-
-	// Set up the HTTP client
-	const ExternalArchiveTimeout = 10 * time.Second
-	client := &http.Client{
-		Transport: &http.Transport{
-			DialContext: (&net.Dialer{
-				Timeout: ExternalArchiveTimeout, // Connection timeout
-			}).DialContext,
-			ResponseHeaderTimeout: ExternalArchiveTimeout, // Read timeout for headers
-		},
-		Timeout: ExternalArchiveTimeout, // Overall request timeout
-	}
-
-	// Get block bytes
-	url := baseUrl + fmt.Sprint(rnd)
-	resp, err := client.Get(url)
-	if err != nil {
-		return EncodedBlockCert{}, fmt.Errorf("failed to download block from external archive: %w", err)
-	}
-	defer resp.Body.Close()
-
-	// Read response body
-	encodedBlockCertBytes, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return EncodedBlockCert{}, fmt.Errorf("failed to read HTTP response body from external archive: %w", err)
-	}
-
-	// Decode block bytes from msgpack
-	{
-		codecHandle := makeCodecHandle()
-
-		dec := codec.NewDecoderBytes(encodedBlockCertBytes, codecHandle)
-		err = dec.Decode(&blk)
-		if err != nil {
-			return EncodedBlockCert{}, fmt.Errorf("failed to decode EncodedBlockCert from msgpack: %w", err)
-		}
-	}
-
-	return blk, nil
-}
-
-func makeCodecHandle() *codec.MsgpackHandle {
-
-	codecHandle := new(codec.MsgpackHandle)
-
-	codecHandle.ErrorIfNoField = true
-	codecHandle.ErrorIfNoArrayExpand = true
-	codecHandle.Canonical = true
-	codecHandle.RecursiveEmptyCheck = true
-	codecHandle.WriteExt = true
-	codecHandle.PositiveIntUnsigned = true
-
-	return codecHandle
-}
